removed unneeded pointer receiver functions from ristretto and moved select/run to internal
