+++ internal/util/select.go
@@ -1,10 +1,10 @@
-package dhpsi
+package util
 
 import (
 	"context"
 )
 
-func sel(ctx context.Context, f func() error) error {
+func Sel(ctx context.Context, f func() error) error {
 	var d = make(chan error)
 	go func() {
 		d <- f()
@@ -18,19 +18,7 @@ func sel(ctx context.Context, f func() error) error {
 	}
 }
 
-func _run(f1 func() error, f2 func() error) chan error {
-	var d = make(chan error, 2)
-	go func() {
-		d <- f1()
-	}()
-	go func() {
-		d <- f2()
-	}()
-
-	return d
-}
-
-func run(fs ...func() error) chan error {
+func Sels(fs ...func() error) chan error {
 	var d = make(chan error, len(fs))
 	for _, f := range fs {
 		f := f
