+++ pkg/dhpsi/dhpsi_ristretto.go
@@ -37,22 +37,22 @@ func NewRistretto(t int) (Ristretto, error) {
 	switch t {
 	case RistrettoTypeGR:
 		var key gr.Scalar
-		return &GR{key: key.Rand()}, nil
+		return GR{key: key.Rand()}, nil
 	case RistrettoTypeR255:
 		var key = r255.NewScalar()
 		var uniformBytes = make([]byte, 64)
 		if _, err := rand.Read(uniformBytes); err != nil {
-			log.Fatalf("could not generate uniform bytes to seed r255")
+			log.Fatalf("could not generate uniform bytes for a r255 scalar")
 		}
 		key.FromUniformBytes(uniformBytes)
-		return &R255{key: key}, nil
+		return R255{key: key}, nil
 	default:
 		return nil, fmt.Errorf("unsupported ristretto type %d", t)
 	}
 }
 
 // "github.com/bwesterb/go-ristretto"
-func (g *GR) DeriveMultiply(dst *[EncodedLen]byte, src []byte) {
+func (g GR) DeriveMultiply(dst *[EncodedLen]byte, src []byte) {
 	var p gr.Point
 	// derive
 	p.DeriveDalek(src)
@@ -62,7 +62,7 @@ func (g *GR) DeriveMultiply(dst *[EncodedLen]byte, src []byte) {
 	q.BytesInto(dst)
 }
 
-func (g *GR) Multiply(dst *[EncodedLen]byte, src [EncodedLen]byte) {
+func (g GR) Multiply(dst *[EncodedLen]byte, src [EncodedLen]byte) {
 	// multiply
 	var p gr.Point
 	p.SetBytes(&src)
@@ -71,7 +71,7 @@ func (g *GR) Multiply(dst *[EncodedLen]byte, src [EncodedLen]byte) {
 }
 
 // "github.com/gtank/ristretto255"
-func (r *R255) DeriveMultiply(dst *[EncodedLen]byte, src []byte) {
+func (r R255) DeriveMultiply(dst *[EncodedLen]byte, src []byte) {
 	var p = r255.NewElement()
 	// derive
 	hash := sha512.Sum512(src)
@@ -84,7 +84,7 @@ func (r *R255) DeriveMultiply(dst *[EncodedLen]byte, src []byte) {
 	copy(dst[:], tmp)
 }
 
-func (r *R255) Multiply(dst *[EncodedLen]byte, src [EncodedLen]byte) {
+func (r R255) Multiply(dst *[EncodedLen]byte, src [EncodedLen]byte) {
 	// multiply
 	var p = r255.NewElement()
 	p.Decode(src[:])
